chore: Move logic types to types dir
