+++ shared/types/index.ts
@@ -9,3 +9,4 @@ export * from './response'
 export * from './submission'
 export * from './user'
 export * from './utils'
+export * from './logic'
